+++ rotors_control/CHANGELOG.rst
@@ -2,7 +2,18 @@
 Changelog for package rotors_control
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+6.0.4 (2020-04-14)
+------------------
+* Improvements in the include folder
+* Add INDI and Mellinger's controllers
+* Contributors: Ria Sonecha, Giuseppe Silano
+
 6.0.3 (2020-03-22)
+------------------
+* Add data saving feature
+* Contributors: Giuseppe Silano
+
+6.0.2 (2020-01-18)
 ------------------
 * Add data saving feature
 * Contributors: Giuseppe Silano